requirements.txt
asgiref==3.8.1
Django==5.0.4
django-extensions==3.2.3
djangorestframework==3.15.1
flake8==7.0.0
mccabe==0.7.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.2.0
python-dotenv==1.0.1
sqlparse==0.5.0
